size_T fixed and .h file

+++ ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: matoledo <matoledo@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/09 11:44:14 by matoledo          #+#    #+#             */
-/*   Updated: 2025/04/09 13:09:02 by matoledo         ###   ########.fr       */
+/*   Updated: 2025/04/09 20:32:11 by matoledo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, const void *src, unsigned long n)
+#include "libft.h"
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	const char	*pt_src;
 	char		*pt_dest;
